% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_brain}
\alias{make_brain}
\title{Make brain data}
\usage{
make_brain(data = "whole_brain_coords_S1.csv")
}
\arguments{
\item{data}{Raw data set of .csv filetype. Defaults to 'whole_brain_coords_S1.csv' from the raw_data folder}
}
\value{
returns six .Rda files named whole_brain_coords_S1, tel_coords_S1, dien_coords_S1, mes_coords_S1, cere_coords_S1, medob_coords_S1 in the data folder
}
\description{
This function imports lizard and snake brain coordinate data (x,y,z), removes snakes, converts species ID's to all lowercase, converts 2D array into 3D array, creates separate data arrays for each brain region of interest, and saves the file as .Rda
}
\details{
the 3D array is [a, b, c]

where

a = landmark id (e.g., landmark 1, landmark 2, landmark 3, etc.) with a max of 61 values

b = the 3D coordinates (e.g., x coord, y coord, z coord) with a max of 3 values

c = species id coded by position in dataframe (e.g., Actonias_meleagris is 1)

species id       #    , , 1

3D coordinates   #    [,1]     [,2]     [,3]

landmark 1       #    [1,] 9104.361 14396.24 7085.975

landmark 2       #    [2,] 9740.752 13144.82 6818.917

Which landmarks ids (1 through 61) correspond to what major brain region

*3 and 45 are shared between Diencephalon, Mesencephalon, and Medulla Oblongata

(even though 3 isn't duplicated between them in their data - weird)

Telencephalon

1:5, 10:14, 19:26

Anterior-most extent of the olfactory bulb 1-2

Lateral-most extent of the olfactory bulb 3-4

Diencephalon

6, 15, 43:45, 54:55, 61

Optic chiasm - mid-sagittal plane 20

Mesencephalon

7:8, 16:17, 27:28, 45, 48:49, 52:53, 58:59

Cerebellum

30:42, 50:51

Medulla oblongota

9, 18, 29, 45:47, 56:57, 60
}
