% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_links.R
\name{make_links}
\alias{make_links}
\title{Make links}
\usage{
make_links()
}
\arguments{
\item{data}{whole_brain_coords_S1.Rda from the data folder}
}
\value{
returns .Rda file named whole_brain_links in the data folder
}
\description{
This isn't a function to run. But I'm creating the links between the different 3D brain coordinates so that we can use them to visualize the brain with wire frames. I originally had this in the R markdown, but that doesn't work because the links function is a active function where I'd have to recreate the links each time.
}
