% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_trait}
\alias{make_trait}
\title{Make trait data}
\usage{
make_trait(data = "lizard_trait_data.csv")
}
\arguments{
\item{data}{Raw data set of .csv filetype. Defaults to 'lizard_trait_data.csv' from the raw_data folder}
}
\value{
returns .Rda file named lizard_trait_data in the data folder
}
\description{
This function imports lizard trait data, converts species ID's to all lowercase, imputes missing continuous data for species, and saves the file as .Rda
}
