% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_tree}
\alias{make_tree}
\title{Make tree data}
\usage{
make_tree(data = "lizard_tree.nex")
}
\arguments{
\item{data}{Raw data set of .nex filetype. Defaults to 'lizard_tree.nex' from the raw_data folder}
}
\value{
returns .Rda file named lzrd_tree in the data folder
}
\description{
This function imports a phylogeny (.nex), converts species ID's to all lowercase, and saves the file as .Rda
}
