% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgls_auto.R
\name{pgls_size}
\alias{pgls_size}
\title{Phylogenetic ANOVA/Regression for Procrustes Shape Variables - SIZE}
\usage{
pgls_size(brain, tree, traitdata, region, outputname)
}
\arguments{
\item{brain}{3D array of brain coordinates}

\item{tree}{tree in nexus format}

\item{traitdata}{dataframe with traits of interest}

\item{region}{string identifying the region of the brain coordinates (e.g., 'Whole Brain')}

\item{outputname}{string indicating the name of the file in a .rtf format (e.g., 'wholebrainshape.rft')}
}
\value{
returns .rtf file containing a gt table into the vignettes folder
}
\description{
Phylogenetic ANOVA/Regression for Procrustes Shape Variables - SIZE
}
\details{
brain, tree, and traitdata need to be organized with the species in the same order and have the same exact name format (for example the tree can't have upper case while the trait has lower case)
}
